update terminology
This set of changes replaces every reference to GNU/Linux with a
reference to Linux in recognition of the fact that not all independent
Linux-based distributions build over the GNU ecosystem. For example,
Alpine Linux uses musl instead of glibc and busybox instead of the GNU
core utilities. Chimera Linux (not yet supported) combines musl and
FreeBSD-based utilities. Void Linux provides official support for musl.