support Chimera Linux images
Chimera Linux entered the alpha stage officially on 2023-06-11. Chimera
Linux combines APK, FreeBSD-based user-space utilities and musl. This
relatively small set of changes demonstrates the ease with which new yet
divergent Linux-based distros can be added to this program.

Compilation and execution of Turret on Chimera Linux is currently out of
scope because some dependencies are not yet available as packages.