improve user specification semantics
Try to reconcile the facts that 0 is the default value of the uint type
in Go as well as the value of the root UID in Linux without fixing an
arbitrary nonzero default value. Thus, make it possible to delegate the
choice of UID to the user-space utility responsible for creating the
user. In addition, disallow the omission of the `user.name` field in the
spec format in favor of a more explicit style.

BREAKING CHANGE: change `user.uid` to `user.id` in the spec format
BREAKING CHANGE: remove `user.create` from the spec format
BREAKING CHANGE: remove `builder.DefaultUser` and `builder.NewSpec`