builder: extract pkg. mgmt. logic from TurretBuilder
Introduce the `TurretPackageManagerInterface` type to represent a
package management interface over Buildah, and the
`TurretPackageManager` type as the interface's canonical implementation.
`TurretPackageManager` executes commands generated by the embedded
`packagemanager.CommandFactory` in the working container.

Thus, we can implement operations that depend on the execution of
multiple commands in a specific order for a particular package manager.
This is necessary for some APT workflows.