builder: extract user and group management data
Introduces the `TurretUserManagerInterface` type to represent a user
and group management interface over Buildah. Implementors of this
interface are expected to consume data generated by the command
factories in the new `usrgrp` package. Moreover, removes the last
remaining distro-specific implementation of `TurretBuilderInterface`,
retaining the interface in the event that it is needed again before
development enters the alpha stage.