rename `LoginShell` to the more concise `Shell`
Indicate that we are not using the provided shell as a login shell.

We prefer to set the specified shell as the command to execute when
starting the container. Thus, we can run the shell as our unprivileged
user. On the other hand, the `login` command is not available in every
base image and may have to be run as the container's root.

We continue to set the login shell during user creation to communicate
an intentionally interactive container experience and facilitate later
experimentation.

BREAKING CHANGE: rename `user.login-shell` to `user.shell` in the spec format
BREAKING CHANGE: rename `LoginShell` field of `spec.User` to `Shell`
BREAKING CHANGE: rename `LoginShell` field of `usrgrp.CreateUserOptions` to `Shell`